Fixed timer bug again

+++ CoolNamePending/Assets/Scripts/LevelSelect.cs
@@ -147,7 +147,7 @@ public class LevelSelect : MonoBehaviour
         else
         {
             string min = ((int)dayTimerCounter / 60).ToString("00");
-            string sec = ((int)dayTimerCounter % 60).ToString("00");
+            string sec = (dayTimerCounter % 60).ToString("00");
             timerText.text = min + ":" + sec;
 
         }
@@ -157,7 +157,7 @@ public class LevelSelect : MonoBehaviour
     IEnumerator updateNightTimer()
     {
         string min = ((int)startTime / 60).ToString("00");
-        string sec = (startTime % 60).ToString("00");
+        string sec = ((int) startTime % 60).ToString("00");
         timerText.text = min + ":" + sec;
         startTime += Time.deltaTime;
         yield return null;
